processor: add skip member check to vector unit
To allow checking of specific index values for being masked or otherwise
skipped in vector operations. This has similar functionality to the
macro `VI_ELEMENT_SKIP` but doesn't rely on being invoked in a loop that
uses i as the index value.

This will allow for use cases such as checking for index `i*j` in nested
loops.

Use the new functionality from the existing macros to consolidate the
code.

Signed-off-by: Dylan Reid <dgreid@rivosinc.com>